+++ src/app/app.module.ts
@@ -7,8 +7,8 @@ import { StoreModule } from '@ngrx/store';
 import * as fromApp from './store/app.reducer';
 import { MaterialModule } from './material.module';
 import { HomeModule } from './components/home/home.module';
-import { LoginModule } from './components/login/login.module';
-import { RegisterModule } from './components/register/register.module';
+import { LoginModule } from './components/auth/login/login.module';
+import { RegisterModule } from './components/auth/register/register.module';
 import { NotFoundModule } from './components/not-found/not-found.module';
 import { AppComponent } from './app.component';
 import { AppRoutingModule } from './app-routing.module';
