+++ src/app/components/auth/register/register.component.ts
