+++ src/app/components/auth/register/register.module.ts
@@ -1,0 +1,20 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { RouterModule } from '@angular/router';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+
+import { RegisterComponent } from './register.component';
+import { MaterialModule } from '../../../material.module';
+
+@NgModule({
+    declarations: [RegisterComponent],
+    imports: [
+        CommonModule,
+        FormsModule,
+        ReactiveFormsModule,
+        MaterialModule,
+        RouterModule,
+    ],
+    exports: [RegisterComponent],
+})
+export class RegisterModule {}
