+++ src/app/components/not-found/not-found.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-
 import {RouterModule} from '@angular/router';
+
 import { NotFoundComponent } from './not-found.component';
 
 @NgModule({
